reg-off: Make officers table work with Manage Off
Make officers table on the database work with the new
"Manage Officers" page on the website.

The tables fields have been changed to allow Null as a value.
This will serve as a placeholder when there is not officer in
place. The value will appear as "-" in the website page.

The primary key has also been set to the Position field now.